Refactor: Refactor the Reservas microservice and add a Lib microservice for RabbitMQ
