+++ reservas/src/main/java/com/hotelalura/RabbitMQ/Service/RabbitMQService.java
@@ -1,0 +1,16 @@
+package com.hotelalura.RabbitMQ.Service;
+
+import org.springframework.amqp.rabbit.core.RabbitTemplate;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+@Service
+public class RabbitMQService {
+
+    @Autowired
+    private RabbitTemplate rabbitTemplate;
+
+    public void sendMessage(String queueName, Object message) {
+        this.rabbitTemplate.convertAndSend(queueName, message);
+    }
+}
